Record what the sweep's entries turned out to be, and drop one dtype literal
`OPEN.md` marks the sixteen closed entries with what each fix actually was, and
says plainly which two were not what the sweep predicted: A8 is a measured null
(JAX's complex `abs` is differentiable at exactly zero), and A3's guard has to
ask `gamma_storage_is_consumable` rather than the input's `gamma_only` flag,
since an ultrasoft gamma run has already been substituted to the whole sphere.

It also records what this machine could not check: the vendored QE tree is
absent here, so `test_input_sweep.py` and the three Kramers tests skip, and B2's
partition was written from `opt_tetra_weights_only`'s description rather than
transcribed from it.

From the trap-check pass over both commits: `_weighted_structure_factors` had a
`dtype=float` literal. It is the only finding; the per-atom structure factor is
checked against a brute-force `sum_a w_a exp(-iG.tau_a)` at 0.0, and reduces to
`structure_factors` exactly when the weights are all one.

+++ defumat/pseudo/potentials.py
@@ -65,7 +65,10 @@ def _weighted_structure_factors(structure, cell, gvectors, weights) -> jnp.ndarr
     this with the same per-species radial transforms and the result is
     ``sum_a w_a rho^at_{t(a)}(G) e^{-iG.tau_a}``.
     """
-    membership = _membership(structure) * jnp.asarray(weights, dtype=float)[None, :]
+    # No dtype literal: ``_membership`` is the real mask and the product follows
+    # it, so the phase sum's precision stays the one the G vectors and the
+    # positions set rather than one written down here.
+    membership = _membership(structure) * jnp.asarray(weights)[None, :]
     return _structure_factors_at(
         gvectors.cartesian(cell), structure.positions, membership
     )
